% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-httptest2.R
\name{use_httptest2}
\alias{use_httptest2}
\title{Use 'httptest2' in your tests}
\usage{
use_httptest2(path = ".")
}
\arguments{
\item{path}{character path to the package}
}
\value{
Nothing: called for file system side effects.
}
\description{
This function adds \code{httptest2} to Suggests in the package DESCRIPTION and
loads it in \code{tests/testthat/setup.R}. Call it once when you're setting up
a new package test suite.
}
\details{
The function is idempotent: if \code{httptest2} is already added to these files,
no additional changes will be made.
}
